General Error/Warning fixing.
[Fix] SceneManager.cpp, include was still from ImGui/Gui.h instead of GUI.h.
[Fix] Renderer.h, m_CurrentFrame and m_PreviousFrame are now int's instead of size_t.

+++ VulkanRenderer/VulkanRendererWindows/MainRenderer/Renderer.h
@@ -97,8 +97,8 @@ private:
 
 	std::vector<FrameData> m_FrameData;
 
-	size_t m_CurrentFrame = 0;
-	size_t m_PreviousFrame = -1;
+	int m_CurrentFrame = 0;
+	int m_PreviousFrame = -1;
 
 	//Primary Vulkan Data;
 	VkInstance mvk_Instance;
